REFACTOR: contract abi & address methods
