+++ src/StakingClient.ts
@@ -11,28 +11,6 @@ export class StakingClient extends Account {
     super(params);
     this.id = contractId;
   }
-
-  /**
-   * Returns the contract address
-   * @param params An object with key vault that indicates the contract whose address is to be retrieved
-   * @returns A formatted address of the specified contract as a string
-   */
-   async getContractAddress(params: { contractId: number }): Promise<string> {
-    await this.initialiseReachAccount();
-    const ctc = this.reachAccount.contract(backend, params.contractId);
-    const contractAddress = await ctc.getContractAddress();
-    return this.reachStdLib.formatAddress(contractAddress);
-  }
-
-  /**
-   * 
-   * @param params contractId which indicates the contract we want to interact with
-   */
-  async getContractAbi(params: { contractId: number }): Promise<AbiInterface> {
-    await this.initialiseReachAccount();
-    const ctc = this.reachAccount.contract(backend, params.contractId);
-    return await ctc.getABI();
-  }
   
   /**
    * @description Get global state of contract
